chore(tabcordion): refactor to allow animation override
