+++ components/tabcordion/src/Tab.js
@@ -1,12 +1,8 @@
 /** @jsx jsx */
 
 import { useState, forwardRef, useEffect } from 'react';
-import { jsx, useBrand, overrideReconciler, getLabel } from '@westpac/core';
+import { jsx, useBrand, overrideReconciler } from '@westpac/core';
 import PropTypes from 'prop-types';
-import { useSpring, animated } from 'react-spring';
-import ResizeObserver from 'resize-observer-polyfill';
-import useMeasure from 'react-use-measure';
-import BezierEasing from 'bezier-easing';
 
 import { defaultItem } from './overrides/item';
 import { defaultAccordionButton } from './overrides/accordionButton';
@@ -14,7 +10,6 @@ import { defaultAccordionButtonIcon } from './overrides/accordionButtonIcon';
 import { defaultPanel } from './overrides/panel';
 import { defaultPanelBody } from './overrides/panelBody';
 import { useTabcordionContext } from './Tabcordion';
-import { usePrevious } from './_utils';
 import pkg from '../package.json';
 
 // ==============================
@@ -54,7 +49,7 @@ export const Tab = forwardRef(
 		const [hidden, setHidden] = useState(!selected);
 		const [initial, setInitial] = useState(true);
 		const [closed, setClosed] = useState(true);
-		const [measureRef, { height: panelBodyHeight }] = useMeasure({ polyfill: ResizeObserver });
+		const [panelBodyHeight, setPanelBodyHeight] = useState(null);
 
 		const defaultOverrides = {
 			Item: defaultItem,
@@ -84,6 +79,9 @@ export const Tab = forwardRef(
 			initial,
 			idx,
 			closed,
+			setClosed,
+			panelBodyHeight,
+			setPanelBodyHeight,
 			context: context.state,
 			overrides: componentOverrides,
 			...rest,
@@ -115,57 +113,6 @@ export const Tab = forwardRef(
 			setHidden(!selected);
 		}, [mode]);
 
-		const prevSelected = usePrevious(selected);
-		const prevHidden = usePrevious(hidden);
-
-		const AnimatedPanel = animated(Panel);
-
-		const animate = useSpring({
-			config: {
-				duration: 300,
-				easing: BezierEasing(0.25, 0.1, 0.25, 1.0), //~'ease' CSS transition timing function
-			},
-			...(mode === 'accordion' && {
-				height: hidden ? 0 : panelBodyHeight,
-			}),
-			from: {
-				height: '', //reset
-			},
-			immediate: initial,
-			onStart: () => {
-				setClosed(false);
-
-				if (mode === 'tabs') {
-					if (selected && !prevSelected) {
-						onOpening({ idx, tabId });
-					} else if (!selected && prevSelected) {
-						onClosing({ idx, tabId });
-					}
-				} else if (mode === 'accordion') {
-					if (!hidden && prevHidden) {
-						// Opening
-						onOpening({ idx, tabId });
-					} else if (hidden && !prevHidden) {
-						// Closing
-						onClosing({ idx, tabId });
-					}
-				}
-			},
-			onRest: () => {
-				if (mode === 'accordion') {
-					if (hidden) {
-						// Closed
-						setClosed(true);
-						onClose({ idx, tabId });
-					} else if (!hidden) {
-						// Opened
-						setClosed(false);
-						onOpen({ idx, tabId });
-					}
-				}
-			},
-		});
-
 		return (
 			<Item state={state} {...itemAttributes(state)} css={itemStyles(state)}>
 				{mode === 'accordion' && (
@@ -184,22 +131,11 @@ export const Tab = forwardRef(
 					</AccordionButton>
 				)}
 
-				<AnimatedPanel
-					ref={ref}
-					style={animate}
-					state={state}
-					{...panelAttributes(state)}
-					css={panelStyles(state)}
-				>
-					<PanelBody
-						ref={measureRef}
-						state={state}
-						{...panelBodyAttributes(state)}
-						css={panelBodyStyles(state)}
-					>
+				<Panel ref={ref} state={state} {...panelAttributes(state)} css={panelStyles(state)}>
+					<PanelBody state={state} {...panelBodyAttributes(state)} css={panelBodyStyles(state)}>
 						{children}
 					</PanelBody>
-				</AnimatedPanel>
+				</Panel>
 			</Item>
 		);
 	}
