+++ components/tabcordion/src/overrides/panelBody.js
@@ -4,12 +4,19 @@ import { forwardRef } from 'react';
 import { jsx, getLabel, formatClassName } from '@westpac/core';
 import { Body } from '@westpac/body';
 import { useSpring, animated } from 'react-spring';
+import ResizeObserver from 'resize-observer-polyfill';
+import useMeasure from 'react-use-measure';
+import mergeRefs from 'react-merge-refs';
 
 // ==============================
 // Component
 // ==============================
 
-const PanelBody = forwardRef(({ state: { mode, selected }, ...rest }, ref) => {
+const PanelBody = forwardRef(({ state: { mode, selected, setPanelBodyHeight }, ...rest }, ref) => {
+	const [measureRef, { height: panelBodyHeight }] = useMeasure({ polyfill: ResizeObserver });
+
+	setPanelBodyHeight(panelBodyHeight);
+
 	const AnimatedBody = animated(Body);
 
 	const fade = useSpring({
@@ -24,7 +31,7 @@ const PanelBody = forwardRef(({ state: { mode, selected }, ...rest }, ref) => {
 		},
 	});
 
-	return <AnimatedBody ref={ref} style={fade} {...rest} />;
+	return <AnimatedBody ref={mergeRefs([measureRef, ref])} style={fade} {...rest} />;
 });
 
 const BlenderPanelBody = ({ className, ...rest }) => (
